+++ toy-rpc/src/server/pubsub/mod.rs
@@ -1,32 +1,27 @@
 //! PubSub impl on the server side
 
-use flume::r#async::{RecvStream, SendSink};
 use flume::{Receiver, Sender};
-use futures::{Sink, Stream};
-use pin_project::pin_project;
 use std::collections::{BTreeMap, BTreeSet, HashMap};
 use std::marker::PhantomData;
-use std::pin::Pin;
 use std::sync::atomic::Ordering;
 use std::sync::Arc;
-use std::task::{Context, Poll};
 use std::time::Duration;
 
 #[cfg(feature = "http_actix_web")]
 use actix::Recipient;
 
-use crate::codec::{Marshal, Reserved, Unmarshal};
-use crate::error::Error;
 use crate::message::{AtomicMessageId, MessageId};
-use crate::pubsub::{AckModeAuto, SeqId, Topic};
+use crate::pubsub::{AckModeAuto, AckModeNone, SeqId};
 
-use super::RESERVED_CLIENT_ID;
-use super::{broker::ServerBrokerItem, ClientId, Server};
+use super::{broker::ServerBrokerItem, ClientId};
 
 #[cfg(all(feature = "async_std_runtime", not(feature = "tokio_runtime")))]
 use async_std::task;
 #[cfg(all(feature = "tokio_runtime", not(feature = "async_std_runtime")))]
 use tokio::task;
+
+pub mod publisher;
+pub mod subscriber;
 
 pub(crate) enum PubSubResponder {
     #[cfg(not(feature = "http_actix_web"))]
@@ -49,6 +44,9 @@ pub(crate) enum PubSubItem {
         topic: String,
         content: Arc<Vec<u8>>,
     },
+    RemovePendingAcks {
+        seq_id: SeqId,
+    },
     Subscribe {
         client_id: ClientId,
         topic: String,
@@ -59,7 +57,8 @@ pub(crate) enum PubSubItem {
         topic: String,
     },
     Ack {
-        seq_id: SeqId
+        seq_id: SeqId,
+        client_id: ClientId,
     },
     Stop,
 }
@@ -70,7 +69,7 @@ pub(crate) struct PubSubBroker<AckMode> {
     listener: Receiver<PubSubItem>,
     subscriptions: HashMap<String, BTreeMap<ClientId, PubSubResponder>>,
     pending_acks: BTreeMap<SeqId, Sender<ClientId>>,
-    pub pub_retry_timeout: Duration,
+    pub_retry_timeout: Duration,
     ack_mode: PhantomData<AckMode>
 }
 
@@ -166,8 +165,9 @@ impl<AckMode: Send + 'static> PubSubBroker<AckMode> {
         }
     }
 
-    pub fn handle_ack(&mut self, seq_id: SeqId) {
-        log::debug!("Received Ack for seq_id: {:?}", &seq_id);
+    pub fn handle_ack(&mut self, seq_id: SeqId, client_id: ClientId) {
+        log::debug!("Received Ack for seq_id: {:?} from client {:?}", &seq_id, &client_id);
+
     }   
 }
 
@@ -217,18 +217,21 @@ impl PubSubBroker<AckModeAuto> {
                     }
                 };
 
-                if let Err(err) = timeout(duration, fut).await {
-                    log::error!("{:?}", err);
-                    let msg = PubSubItem::PublishRetry {
-                        client_ids: set, 
-                        seq_id, 
-                        topic: topic, 
-                        content
-                    };
-                    pubsub_tx.send_async(msg)
+                let msg = match timeout(duration, fut).await {
+                    Ok(_) => PubSubItem::RemovePendingAcks { seq_id },
+                    Err(err) => {
+                        log::error!("{:?}", err);
+                        PubSubItem::PublishRetry {
+                            client_ids: set, 
+                            seq_id, 
+                            topic: topic, 
+                            content
+                        }
+                    }
+                };
+                pubsub_tx.send_async(msg)
                         .await
-                        .unwrap_or_else(|_| log::error!("Error found sending PubSubItem::PublishRetry"))
-                }
+                        .unwrap_or_else(|err| log::error!("{:?}", err))
             });
         }
     }
@@ -281,6 +284,9 @@ macro_rules! impl_pubsub_broker_for_ack_modes {
                             } => {
                                 self.handle_publish_retry(client_ids, seq_id, topic, content).await
                             },
+                            PubSubItem::RemovePendingAcks{ seq_id } => {
+                                self.pending_acks.remove(&seq_id);
+                            },
                             PubSubItem::Subscribe {
                                 client_id,
                                 topic,
@@ -291,8 +297,8 @@ macro_rules! impl_pubsub_broker_for_ack_modes {
                             PubSubItem::Unsubscribe { client_id, topic } => {
                                 self.handle_unsubscribe(client_id, topic)
                             },
-                            PubSubItem::Ack{seq_id} => {
-                                self.handle_ack(seq_id)
+                            PubSubItem::Ack{seq_id, client_id} => {
+                                self.handle_ack(seq_id, client_id)
                             },
                             PubSubItem::Stop => return,
                         }
@@ -332,181 +338,38 @@ fn send_broker_item_publication(
     true
 }
 
-/* -------------------------------------------------------------------------- */
-/*                                 Public API                                 */
-/* -------------------------------------------------------------------------- */
+// /* -------------------------------------------------------------------------- */
+// /*                                 Public API                                 */
+// /* -------------------------------------------------------------------------- */
 
-/// Publisher on the server side
-#[pin_project]
-pub struct Publisher<T: Topic, C: Marshal> {
-    #[pin]
-    inner: SendSink<'static, PubSubItem>,
-    counter: AtomicMessageId,
-    marker: PhantomData<T>,
-    codec: PhantomData<C>,
-}
-
-impl<T: Topic, C: Marshal> From<Sender<PubSubItem>> for Publisher<T, C> {
-    fn from(inner: Sender<PubSubItem>) -> Self {
-        Self {
-            inner: inner.into_sink(),
-            counter: AtomicMessageId::new(0),
-            marker: PhantomData,
-            codec: PhantomData,
-        }
-    }
-}
-
-impl<T: Topic, C: Marshal> Sink<T::Item> for Publisher<T, C> {
-    type Error = Error;
-
-    fn poll_ready(
-        self: Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> Poll<Result<(), Self::Error>> {
-        let this = self.project();
-        this.inner.poll_ready(cx).map_err(|err| err.into())
-    }
-
-    fn start_send(self: Pin<&mut Self>, item: T::Item) -> Result<(), Self::Error> {
-        let this = self.project();
-        let topic = T::topic();
-        let msg_id = this.counter.fetch_add(1, Ordering::Relaxed);
-        let body = C::marshal(&item)?;
-        let content = Arc::new(body);
-        let item = PubSubItem::Publish {
-            client_id: RESERVED_CLIENT_ID,
-            msg_id,
-            topic,
-            content,
-        };
-        this.inner.start_send(item).map_err(|err| err.into())
-    }
-
-    fn poll_flush(
-        self: Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> Poll<Result<(), Self::Error>> {
-        let this = self.project();
-        this.inner.poll_flush(cx).map_err(|err| err.into())
-    }
-
-    fn poll_close(
-        self: Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> Poll<Result<(), Self::Error>> {
-        let this = self.project();
-        this.inner.poll_close(cx).map_err(|err| err.into())
-    }
-}
-
-/// Subscriber on the client side
-#[pin_project]
-pub struct Subscriber<T: Topic, C: Unmarshal> {
-    #[pin]
-    inner: RecvStream<'static, ServerBrokerItem>,
-    topic: String,
-    marker: PhantomData<T>,
-    codec: PhantomData<C>,
-}
-
-impl<T: Topic, C: Unmarshal> From<Receiver<ServerBrokerItem>> for Subscriber<T, C> {
-    fn from(inner: Receiver<ServerBrokerItem>) -> Self {
-        Self {
-            inner: inner.into_stream(),
-            topic: T::topic(),
-            marker: PhantomData,
-            codec: PhantomData,
-        }
-    }
-}
-
-impl<T: Topic, C: Unmarshal> Stream for Subscriber<T, C> {
-    type Item = Result<T::Item, Error>;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let this = self.project();
-        match this.inner.poll_next(cx) {
-            Poll::Pending => Poll::Pending,
-            Poll::Ready(opt) => match opt {
-                Some(item) => match item {
-                    ServerBrokerItem::Publication {
-                        seq_id: _,
-                        topic,
-                        content,
-                    } => {
-                        let result = match &topic == this.topic {
-                            true => C::unmarshal(&content),
-                            false => Err(Error::Internal("Mismatched topic".into())),
-                        };
-                        Poll::Ready(Some(result))
-                    }
-                    _ => {
-                        let result = Err(Error::Internal("Invalid PubSub item".into()));
-                        Poll::Ready(Some(result))
-                    }
-                },
-                None => Poll::Ready(None),
-            },
-        }
-    }
-}
-
-cfg_if::cfg_if! {
-    if #[cfg(any(
-        any(feature = "docs", doc),
-        all(
-            feature = "serde_bincode",
-            not(feature = "serde_json"),
-            not(feature = "serde_cbor"),
-            not(feature = "serde_rmp"),
-        ),
-        all(
-            feature = "serde_cbor",
-            not(feature = "serde_json"),
-            not(feature = "serde_bincode"),
-            not(feature = "serde_rmp"),
-        ),
-        all(
-            feature = "serde_json",
-            not(feature = "serde_bincode"),
-            not(feature = "serde_cbor"),
-            not(feature = "serde_rmp"),
-        ),
-        all(
-            feature = "serde_rmp",
-            not(feature = "serde_cbor"),
-            not(feature = "serde_json"),
-            not(feature = "serde_bincode"),
-        ),
-    ))] {
-        use crate::codec::DefaultCodec;
-        use crate::pubsub::AckModeNone;
-
-        type PhantomCodec = DefaultCodec<Reserved, Reserved, Reserved>;
-
-        impl Server<AckModeNone> {
-            /// Creates a new publihser on a topic
-            pub fn publisher<T: Topic>(&self) -> Publisher<T, PhantomCodec> {
-                let tx = self.pubsub_tx.clone();
-                Publisher::from(tx)
-            }
-
-            /// Creates a new subscriber on a topic
-            ///
-            /// Multiple subscribers can be created on the server side
-            #[cfg(not(feature = "http_actix_web"))]
-            #[cfg_attr(feature = "docs", doc(cfg(not(feature = "http_actix_web"))))]
-            pub fn subscriber<T: Topic>(&self, cap: usize) -> Result<Subscriber<T, PhantomCodec>, Error> {
-                let (sender, rx) = flume::bounded(cap);
-                let client_id = RESERVED_CLIENT_ID;
-                let topic = T::topic();
-                let sender = PubSubResponder::Sender(sender);
-                self.pubsub_tx.send(PubSubItem::Subscribe{client_id, topic, sender})?;
-                Ok(
-                    Subscriber::from(rx)
-                )
-            }
-        }
-    }
-}
+// cfg_if::cfg_if! {
+//     if #[cfg(any(
+//         any(feature = "docs", doc),
+//         all(
+//             feature = "serde_bincode",
+//             not(feature = "serde_json"),
+//             not(feature = "serde_cbor"),
+//             not(feature = "serde_rmp"),
+//         ),
+//         all(
+//             feature = "serde_cbor",
+//             not(feature = "serde_json"),
+//             not(feature = "serde_bincode"),
+//             not(feature = "serde_rmp"),
+//         ),
+//         all(
+//             feature = "serde_json",
+//             not(feature = "serde_bincode"),
+//             not(feature = "serde_cbor"),
+//             not(feature = "serde_rmp"),
+//         ),
+//         all(
+//             feature = "serde_rmp",
+//             not(feature = "serde_cbor"),
+//             not(feature = "serde_json"),
+//             not(feature = "serde_bincode"),
+//         ),
+//     ))] {
+        
+//     }
+// }
